Module 11: UFO Sightings
The purpose of the Module 11 file was to create an interactive webpage that allows you to input your own parameters to search for historical UFO sightings.  We built our webpage using Javascript and set up parameters which included Date, City, State, Country, and Shape of the UFO.

The origin of over information was a json data file.  Below is a shot of the website created.